Rename Call → SExpr, Array → SqExpr
I don't think either of these syntactical constructs is going to consistently mean either of the things implied by the old names. In particular, I intend square brackets to be used for something other than array literals.

SqExpr is kind of a bad name but it's the best I could think of.