Major refactoring
In order to support macros and the like, the compiler has been majorly refactored.
Much of the parsing logic has been moved out of the parser and into the main compiler (the new `compile` module).
It now uses a coroutine-like method of execution, where compilation of one function can be paused at any time if an unknown symbol is encountered. This allows macros to be handled appropriately.